исправить валидацию на странице профиля
