+++ src/components/Profile/Profile.css
@@ -50,8 +50,6 @@
   flex-direction: column;
 }
 
-
-
 .profile-form__error{
   color: #FF004C;
   font-size: 11px;
@@ -105,6 +103,66 @@
   color: #EE3465;
   font-weight: 500;
   text-decoration: none;
+}
+.profile-input {
+  display: flex;
+  gap: 10px;
+  justify-content: space-between;
+  padding: 0 10px 0;
+  padding-bottom: 16px;
+  padding-top: 16px;
+  width: 100%;
+  font-size: 11px;
+  font-weight: 500;
+  line-height: 13px;
+  border-bottom: 1px solid #E8E8E8;
+  flex-direction: row-reverse;
+}
+
+.profile-input__error-message{
+  font-size: 10px;
+  font-weight: 400;
+  line-height: 12px;
+  color: #EE3465;
+}
+
+.profile-input:last-of-type {
+  border-bottom: none;
+}
+
+.profile-input__caption{
+  color: black;
+  transition: all 0.3s ease;
+}
+
+.profile-input__input {
+  border: 0;
+  padding: 0;
+  font-weight: 400;
+  text-align: right;
+  flex-grow: 1;
+}
+
+.profile-input__input_error {
+  color: #EE3465;
+}
+
+.profile-input__input:disabled{
+  background-color: #FFFFFF;
+  color: black;
+}
+
+.profile-input__input:focus ~ .profile-input__caption{
+  color: #FF6838;
+}
+
+.profile-input__input:focus{
+  background-color: transparent;
+  outline: none;
+}
+
+.profile-input:has(.profile-input__input:focus){
+  background-color: #f5f5f5;
 }
 
 @media screen and (max-width: 768px) {
